Signout Api

+++ admin-app/src/actions/auth.action.js
@@ -78,9 +78,20 @@ export const isUserLoggedIn = () => {
 
 export const signout = () => {
     return async dispatch => {
-        localStorage.clear();
-        dispatch({
-            type: authConstants.LOGOUT_REQUEST
-        })
+        dispatch({ type: authConstants.LOGOUT_REQUEST });
+        const res = await axiosInstance.post(`/admin/signout`);
+        console.log(res.status)
+        if(res.status === 200){
+            localStorage.clear();
+            dispatch({
+            type: authConstants.LOGOUT_SUCCESS
+        });
+        } else {
+            dispatch({
+                type: authConstants.LOGOUT_FAILURE,
+                payload: { error: res.data.error}
+            })
+        }
+        
     }
 }
